Move questions to i18n directory
