+++ server/models/view-models/contactInformationCheckAnswers.ts
@@ -1,5 +1,5 @@
 import paths from '../../constants/paths'
-import questions from '../../constants/questions'
+import questions from '../../i18n/en/questions'
 import crownCourts from '../../reference/crown-courts'
 import magistratesCourts from '../../reference/magistrates-courts'
 import notifyingOrganisations from '../../reference/notifying-organisations'
