+++ server/types/content/questions.ts
@@ -1,0 +1,5 @@
+type Questions = {
+
+}
+
+export default Questions
