+++ server/utils/nunjucksSetup.ts
@@ -12,7 +12,7 @@ import probationRegions from '../reference/probation-regions'
 import youthJusticeServiceRegions from '../reference/youth-justice-service-regions'
 import crownCourts from '../reference/crown-courts'
 import magistratesCourts from '../reference/magistrates-courts'
-import questions from '../constants/questions'
+import questions from '../i18n/en/questions'
 import responsibleOrganisations from '../reference/responsible-organisations'
 import notifyingOrganisations from '../reference/notifying-organisations'
 import sentenceTypes from '../reference/sentence-types'
